+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 Guide: https://keepachangelog.com/en/1.0.0/
 -->
 
+## Unreleased
+
+### Fixed
+- [Place Autocomplete]: request all possible `PlaceType` values in case if types were not specified in a search query options.
+- [Address Autofill]: fixed reverse geocoding query, removed unsupported types from query.
+
 ## 1.0.0-rc.4 - 2023-05-19
 
 ### Added